[#4248] refactor(field-types): relocate to Onetime::Models::FieldTypes
Field types exist solely to serve the Familia horreum models, so house
them with the other model-support code (models/features/*) instead of
adding to the lib/onetime top level. Move lib/onetime/field_types{,/} to
lib/onetime/models/field_types{,/}, nest the classes under
Onetime::Models, and mirror the move in try/ and spec/.
