+++ lib/onetime/models/field_types/integer_field_type.rb
@@ -1,0 +1,40 @@
+# lib/onetime/models/field_types/integer_field_type.rb
+#
+# frozen_string_literal: true
+
+require_relative 'coercing_field_type'
+
+module Onetime
+  module Models
+    module FieldTypes
+      # Coerces numeric inputs to native Integer values for Familia scalar fields.
+      class IntegerFieldType < CoercingFieldType
+        NUMERIC_STRING = /\A\s*-?\d+(?:\.\d+)?\s*\z/
+
+        def initialize(name, storage: :native, **)
+          validate_storage!(name, storage)
+          super(name, **)
+        end
+
+        def coerce(value)
+          return nil if value.nil? || value == ''
+          return value if value.is_a?(Integer)
+          return value.to_i if value.is_a?(Float)
+          return value.to_i if !value.is_a?(String) && value.respond_to?(:to_i)
+          return value.to_i if value.is_a?(String) && NUMERIC_STRING.match?(value)
+
+          log_coercion_issue(value)
+          nil
+        end
+
+        private
+
+        def validate_storage!(name, storage)
+          return if storage == :native
+
+          raise ArgumentError, "Unknown integer storage #{storage.inspect} for field #{name} (expected :native)"
+        end
+      end
+    end
+  end
+end
