+++ try/unit/models/field_types/integer_field_type_try.rb
@@ -1,9 +1,9 @@
-# try/unit/field_types/integer_field_type_try.rb
+# try/unit/models/field_types/integer_field_type_try.rb
 #
 # frozen_string_literal: true
 
-require_relative '../../support/test_helpers'
-require 'onetime/field_types'
+require_relative '../../../support/test_helpers'
+require 'onetime/models/field_types'
 
 OT.boot! :test, false
 
@@ -18,7 +18,7 @@ class IntegerFieldTypeTryToIValue
 end
 
 class IntegerFieldTypeTryModel < ::Familia::Horreum
-  extend Onetime::FieldTypes::Macros
+  extend Onetime::Models::FieldTypes::Macros
 
   prefix :try_integer_field_type
   identifier_field :probeid
